Move around some files and figured out guild command permissions
