+++ src/commands/economy/lotto.config.ts
@@ -1,4 +1,4 @@
-import Command from '../../common.commands.config';
+import Command from '../../utils/common.commands.config';
 import economyServices from '../../services/economy.services';
 import { CommandInteraction, GuildMember, MessageEmbed, TextChannel } from 'discord.js';
 import { Document } from 'mongoose';
